[Coding]: Added Input Validations.
-----------------------------------------------------
Added Validations on UserName EditText to Validate that length is 5 or more characters.

Added Validations on Email EditText to Validate that it is proper Email.

Added Validations on Password EditText to Validate that length is 8 or more characters.

Added Validations on Password EditText to Validate that it is not same as UserName.

Added Validations on RETyped Password EditText to Validate that it matches with entered Password.