[Coding]: Added User Validation.
-----------------------------------------------------
Added Validations on UserName EditText to Validate that entered User name already Exists in database or Not.
If It Already exists then an dialog is shown to notify the user about it and EditText fields are clear so that user can enter new Data.

DBLogin : Added Funtion to check user is present or Not it returns boolean value if user is already registered once.