[Coding] : Added common Funtion and Validate Password.
----------------------------------------------------
Added Funtion isAvailable which checks user is available or not in Database.
It is used on RegistrationActivity to check if user is already Registered.
&
It will also be used to check if user is registered or NOT on Login Page.

Added Function ValidateUser which validates the password if user exists return boolean value.