Implemented Timer functionality.
When clicked on timer button it shows bottom sheet dialog.
This dialog contains 3 option 15 minutes, 30 minutes, and 60 minutes.
when any option is selected the app is scheduled to close in that time.
When the user again clicks on the button again it shows an dialog asking about does he want to cancel the timer.